requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
Flask-QRcode==3.1.0
SQLAlchemy==2.0.27
mysql-connector-python==8.3.0
redis==5.0.1
bcrypt==4.1.2
python-dotenv==1.0.1
sentry-sdk==1.40.0
python-json-logger==2.0.7
gunicorn==21.2.0
Werkzeug==3.0.1
qrcode==7.4.2
pillow==10.2.0
alembic==1.13.1
